requirements.txt
requests==2.31.0
termcolor==2.4.0
pyyaml==6.0.1
flask==3.0.2
flask-socketio==5.3.6
python-socketio==5.11.1
eventlet==0.35.2
pandas==2.2.1
plotly==5.19.0
gunicorn==21.2.0
python-dotenv==1.0.1 